Move target selector extractor method to common module
